requirements.txt
# SBI HyperSBI2 自動売買システム 依存ライブラリ

# HTTP クライアント
requests==2.31.0
httpx==0.27.0

# HTML パース (スクレイピング用)
beautifulsoup4==4.12.3
lxml==5.2.1

# 環境変数管理
python-dotenv==1.0.1

# データ操作
pandas==2.2.2

# スケジューリング
schedule==1.2.2

# ログ管理
loguru==0.7.2

# リトライ処理
tenacity==8.3.0

# 日時処理
python-dateutil==2.9.0

# 型ヒント (Python 3.8 以下対応)
typing-extensions==4.11.0
